SAM: Refactors suggested in PR
