+++ scalafix-tests/output/src/main/scala-2.11/test/ConvertSingleAbstractMethod.scala
