+++ scalafix-tests/unit/src/test/scala/scalafix/tests/rule/SemanticTests.scala
@@ -39,15 +39,11 @@ object SemanticTests {
     )
   )
   def classpath: Classpath = {
-    val dependencies = BuildInfo.dependencyClasspath
-      .map(file => AbsolutePath(file.toString))
-      .toList
-
     Classpath(
       List(
         AbsolutePath(BuildInfo.semanticSbtClasspath),
         AbsolutePath(BuildInfo.semanticClasspath)
-      ) ::: dependencies
+      )
     )
   }
 }
