+++ website/src/main/tut/docs/rules/ConvertSingleAbstractMethod.md
@@ -1,9 +1,9 @@
 ---
 layout: docs
-title: SingleAbstractMethod
+title: ConvertSingleAbstractMethod
 ---
 
-# SingleAbstractMethod
+# ConvertSingleAbstractMethod
 
 http://www.scala-lang.org/files/archive/spec/2.12/06-expressions.html#sam-conversion
